Improve the logging date format.
Also added a TODO to look into adding an inhibitor handler (for adding a
blacklist if need be), rearranged the version check code in start(), and
migrated a couple strings that were too long to a dedicated file called
strings.json. May migrate other strings to that file, but I'm not 100%
sure yet. Depends on the length of the string.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>