Added profile support.
Also added a feature to Last.fm command where if available, it reads
the username in the database and uses that instead of using an argument.
Note however that due to the way this works so far, this feature cannot
be used in conjuction with setting the Recent Tracks song limit.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>