Remove Windows cargo check.
Because I primairly develop on Windows, this check is honestly
redundant. Plus it's pretty much the same result across platforms anyway
so there's not really much point in having three checks doing the same
thing.

Also rename the Workflow to Check Project and rename the Lints check to
Check lints.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>