info: Added about command.
This command shows various information about Ellie, such as memory and
CPU usage, as well as various Discord-related informational statistics
such as users, guilds, and shards.

With this change, comes the removal of the version command. I also
decided to remove the built crate, since there was hardly anything that
was using it.

The bot owner and owner ID is now logged during the ready event
initialization as well, and the if else statement determining whether or
not the bot is in debug mode has been removed, since the built crate is
no longer present.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>