Changes. They are a happenin'.
This reflects the many, many changes I have made to the bot since the
last commit. Please note, there are MANY, MANY changes made. Most of
them are just me adapting the commands in my non-async codebase to this
new one, as well as migrating to a new configuration system and database
system / library (postgres and sqlx), and also finally introducing a
proper voice system based on serenity's new voice implementation crate,
songbird.

Please note that the voice system requires a Lavalink-compatible server
running, and the bot will refuse to start up if you don't have one
active. Andesite or Lavalink itself are recommended, and this requires
Java to be installed, preferably the newest, which is Java 15 at the
time of me writing this, and no, there is not a Lavalink implementation
written in Rust, otherwise I would be recommending the use of it instead
of Java-based ones.

This is not the last commit to this project. I am back, and starting to
become active again, however I will not be as active as I once was until
mid/late December, at which point my current job is ending, which I am
ok with, unless they somehow keep me on, which I will also be 100% okay
with too.

Also, it should be noted that there are still some commands that were in
my non-async codebase that are still not present, e.g. the commands
integrating with Reddit / GitHub, as well as the weather command that
used Dark Sky. Dark Sky is shutting down in mid-2021, so I don't see
much point in bringing back the weather command unless I can find
another weather API to replace Dark Sky.

Anyways, that's all from me for now. Take care to anyone reading this,
and Happy Holidays :christmas_tree:

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>