net_utils: Fixed a search issue in get_album_artwork().
Fairly simple bug fix but took me more time than I would've liked to
fix. I changed the market to Canada, as in some situations, albums do
not share the same name as their US counterpart, e.g. in the situation
of Light's album "Skin & Earth", which is actually "Skin&Earth" in the
United States, for some reason. US & Canada have pretty much the same
musical selection, so this shouldn't really change anything.

However, this bug is still present in the Spotify comamnd suite, but for
that I'll properly look into adding a command flag / argument that lets
the user manually choose the market instead of having aspotify (this is
the library I use for handling any and all requests to the Spotify Web
API) automatically choosing the market.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>