chore: The great removal of error embeds.
These were unnecessary just to display error messages, not to mention
the error messages were a bit too verbose anyways. This makes stuff a
lot cleaner, and pretty much removes the majority of the unnecessary
embeds that were previously implemented. This also has the added plus of
making the codebase a bit cleaner and not as messy (though, to be honest
that is mostly because of rustfmt's excessive urge to split function
parameters for any function it deems to be too long to their own
newline.

For example, for the send_message() function calls, they could be
reduced from their current form, to this, improving readability and
removing the excessive amount of newline bloat that is in the codebase,
majority of which are caused by every function parameter being split
onto its own newline.

message.channel_id.send_message(|message| {
    message.embed(|embed| {
        embed.title("Test Embed");
	embed
  }
}

Oh well. Supposedly rustfmt 2.0.0 has the ability to customize chain
length, so whenever that gets released (whenever it actually does get
released, as Rust isn't 2.0 yet, and it follows a strict semver policy)
I will probably most definitely be enabling that in my configuration for
rustfmt.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>