removal: Removed voice module, for now.
Songbird wasn't playing nice with my serenity version (again), and I
couldn't be bothered to try and fix the existing module. I plan on doing
a rewrite of the voice module soon, to fix and add to the voice / music
module.

I also updated the bot to be based on Tokio 1.x instead of 0.x, finally.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>